# backend/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: United Pass Backend API
  description: |
    United Pass is a unified identity, account, organization, OAuth application,
    authorization, Provider synchronization and audit platform.

    This OpenAPI specification is the machine-readable contract. The human-readable
    detail lives in `frontend/docs/api-contracts.md`. Both must stay synchronized.

    Phase 0 scope: operational endpoints only.
    Phase 1 scope: session establishment, MFA, logout, current user and
    permission capabilities.
    Phase 2 scope: OAuth application and client management plane (scope
    catalog, application/client lifecycle, secret rotation, reauthentication).
    See `../docs/adr-0004.md` for the architecture decisions.

    Business API paths use the full `/api/v1/...` prefix explicitly rather than
    relying on a global server base path, so that root-level operational endpoints
    and versioned API endpoints can coexist accurately.
  version: 0.3.0
  contact:
    name: United Pass backend team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: /
    description: Same-origin production
...
paths:
...
  /api/v1/me:
    get:
      tags: [Account]
      summary: Get the current user
      description: |
        Returns the profile of the authenticated user bound to the `up_session`
        cookie.

        The stable United Pass `userId` is the identity primary key. It is not
        an email, phone number or external Provider identifier, and it does not
        change when an external user becomes an employee.

        `employeeProfile` is optional and present only when the user has an
        active employee persona. When an external user is linked to an employee
        profile, the same `userId` is preserved and the consumer persona
        remains.

        Responses must not be cached.
      operationId: getCurrentUser
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Current user profile.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
            Cache-Control:
              description: Prevents the browser and intermediaries from caching the response.
              schema:
                type: string
                enum: [no-store]
                default: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
              example:
                userId: usr_01HQFGXKN8H2C4JVMW7K3R4D8Z
                displayName: 林知行
                nickname: 知行
                avatarUrl: https://cdn.example.com/avatars/usr_01HQFGXKN8H2C4JVMW7K3R4D8Z.webp
                email: zhixing.lin@example.com
                phoneMasked: "+86 138****0123"
                personas: [consumer, employee]
                employeeProfile:
                  employeeId: EMP-10032
                  departmentName: 平台架构部
                  title: 高级工程师
        "401":
          description: No valid session cookie was supplied.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: auth.unauthenticated
                  message: "请先登录后再继续。"
                  requestId: req_0123456789abcdef
        "500":
          description: Unexpected server failure.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: internal.unexpected
                  message: "服务暂时无法处理该请求。"
                  requestId: req_0123456789abcdef
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: up_session
      description: |
        Browser session cookie (`up_session`) issued by United Pass. The cookie
        is HttpOnly, Secure and carries an explicit SameSite policy. The browser
        must not receive raw authentication-provider Refresh Tokens or
        long-lived tokens.
  headers:
    X-Request-ID:
      description: |
        The request ID used for log, error and audit correlation. The server
        accepts a valid upstream ID (alphanumeric, dash, underscore, colon;
        8-128 chars) or generates a cryptographically random `req_`-prefixed ID.
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_:-]{8,128}$"
        example: req_0123456789abcdef0123456789abcdef
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code
          examples:
            internal: internal.unexpected
            validation: validation.failed
            reauth: session.reauthentication_required
        message:
          type: string
          description: Safe, user-displayable message
          example: "服务暂时无法处理该请求。"
        requestId:
          type: string
          description: Request ID for troubleshooting and audit correlation
          example: req_0123456789abcdef
        fieldErrors:
          type: array
          description: Field-level validation errors (array format)
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: The field that failed validation
          example: redirectUris[0]
        message:
          type: string
          description: Safe, user-displayable field error message
          example: "该重定向地址未登记。"
...
    UserPersona:
      type: string
      enum: [consumer, employee]
      description: |
        A persona attached to a stable United Pass user. An external user who
        becomes an employee keeps the same `userId` and retains the consumer
        persona.
    EmployeeProfile:
      type: object
      description: |
        Active employee profile linked to the stable user identity. Optional —
        present only when the user has an active employee persona.
      required: [employeeId, departmentName, title]
      additionalProperties: false
      properties:
        employeeId:
          type: string
          description: Employee number unique within its scope.
          example: EMP-10032
        departmentName:
          type: string
          description: Display name of the employee's department.
          example: 平台架构部
        title:
          type: string
          description: Employee position or title.
          example: 高级工程师
    CurrentUser:
      type: object
      description: |
        Profile of the authenticated user. The `userId` is the stable United
        Pass identity primary key — not an email, phone number or external
        Provider identifier. `employeeProfile` is present only when the user
        has an active employee persona.
      required: [userId, displayName, email, phoneMasked, personas]
      additionalProperties: false
      properties:
        userId:
          type: string
          description: Stable United Pass user ID, used as the OIDC subject.
          example: usr_01HQFGXKN8H2C4JVMW7K3R4D8Z
        displayName:
          type: string
          description: User-facing display name.
          example: 林知行
        nickname:
          type: string
          nullable: true
          description: Optional user-chosen nickname.
          example: 知行
        avatarUrl:
          type: string
          format: uri
          nullable: true
          description: |
            Same-origin or controlled media domain URL of the user's avatar.
            Stored after server-side validation, re-encoding and metadata
            stripping. The frontend never submits an external image URL.
          example: https://cdn.example.com/avatars/usr_01HQFGXKN8H2C4JVMW7K3R4D8Z.webp
        email:
          type: string
          format: email
          description: Verified or primary email address.
          example: zhixing.lin@example.com
        phoneMasked:
          type: string
          description: Masked phone number safe for user display.
          example: "+86 138****0123"
        personas:
          type: array
          description: Personas attached to the user identity.
          items:
            $ref: "#/components/schemas/UserPersona"
          example: [consumer, employee]
        employeeProfile:
          allOf:
            - $ref: "#/components/schemas/EmployeeProfile"
          nullable: true
          description: Active employee profile, present only for employees.
...